% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ece.R
\name{ece}
\alias{ece}
\title{Expected Calibration Error}
\description{
Calculates the Expected Calibration Error (ECE) for classification tasks.
ECE measures the difference between predicted probabilities and actual outcomes,
providing an assessment of how well the predicted probabilities are calibrated.
}
\examples{
# Example usage
set.seed(1)
library(mlr3verse)

# Load the task
data("Sonar", package = "mlbench")
task = as_task_classif(Sonar, target = "Class", positive = "M")
splits = partition(task)
task_train = task$clone()$filter(splits$train)
task_test = task$clone()$filter(splits$test)

# Initialize the base learner
learner_uncal <- lrn("classif.ranger", predict_type = "prob")

# Initialize the calibrated learner
rsmp <- rsmp("cv", folds = 5)
learner_cal <- as_learner(PipeOpCalibrationPerFold$new(learner = learner_uncal,
method = "platt", rsmp = rsmp))

# Set ID's for the learners
learner_cal$id <- "Calibrated Learner"

# Train the calibrated learner
learner_cal$train(task_train)

# Predict the learner
prediction <- learner_cal$predict(task_test)

# Calculate the ECE
ece <- prediction$score(ece$new())

}
\references{
Schwarz J, Heider D (2019). “GUESS: Projecting Machine Learning Scores to Well-Calibrated Probability Estimates for Clinical Decision Making.” \emph{Bioinformatics}, \emph{35}(14), 2458-2465.
}
\section{Super classes}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{\link[mlr3:MeasureClassif]{mlr3::MeasureClassif}} -> \code{ece}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ece-new}{\code{ece$new()}}
\item \href{#method-ece-clone}{\code{ece$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate"><a href='../../mlr3/html/Measure.html#method-Measure-aggregate'><code>mlr3::Measure$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format"><a href='../../mlr3/html/Measure.html#method-Measure-format'><code>mlr3::Measure$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help"><a href='../../mlr3/html/Measure.html#method-Measure-help'><code>mlr3::Measure$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print"><a href='../../mlr3/html/Measure.html#method-Measure-print'><code>mlr3::Measure$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score"><a href='../../mlr3/html/Measure.html#method-Measure-score'><code>mlr3::Measure$score()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ece-new"></a>}}
\if{latex}{\out{\hypertarget{method-ece-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{ece} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ece$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ece-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ece-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ece$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
