% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureClassifConfidenceECE.R
\name{MeasureClassifConfidenceECE}
\alias{MeasureClassifConfidenceECE}
\alias{mlr_measures_classif.conf_ece}
\title{Confidence expected calibration error}
\description{
Confidence ECE of Guo et al. (2017), Eq. 3.
Dictionary key \code{classif.conf_ece}; default \code{bins = 15}.
Scores the maximum predicted probability against whether the predicted label is correct
(labels pooled; unlike \link{MeasureClassifTopLabelECE}, which also conditions on the predicted label).
With \eqn{\hat{c}_i = \arg\max_k p_{ik}}, \eqn{h_i = \max_k p_{ik}}, \eqn{z_i = 1(y_i = \hat{c}_i)}:
\deqn{\mathrm{conf{-}ECE} = \sum_{b = 1}^{B} \frac{|B_b|}{N}\,\left|\mathrm{mean}(z_i : i \in B_b) - \mathrm{mean}(h_i : i \in B_b)\right|}

Fixed-width bins use right-closed intervals \eqn{((b - 1)/B, b/B]}; confidence zero goes in the first bin;
empty bins have zero weight.
Top-probability ties break to the first class in column order.
}
\examples{
if (mlr3misc::require_namespaces("mlr3learners", quietly = TRUE) &&
  mlr3misc::require_namespaces("rpart", quietly = TRUE)) {
  task = mlr3::as_task_classif(iris, target = "Species")
  learner = mlr3::lrn("classif.rpart", predict_type = "prob")
  learner$train(task)
  learner$predict(task)$score(mlr3::msr("classif.conf_ece"))
}

}
\references{
Guo C, Pleiss G, Sun Y, Weinberger KQ (2017).
\dQuote{On Calibration of Modern Neural Networks.}
In \emph{Proceedings of the 34th International Conference on Machine Learning}, volume 70, 1321-1330.
\url{https://proceedings.mlr.press/v70/guo17a.html}.
}
\section{Super classes}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{\link[mlr3:MeasureClassif]{mlr3::MeasureClassif}} -> \code{MeasureClassifConfidenceECE}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeasureClassifConfidenceECE-new}{\code{MeasureClassifConfidenceECE$new()}}
\item \href{#method-MeasureClassifConfidenceECE-clone}{\code{MeasureClassifConfidenceECE$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate"><a href='../../mlr3/html/Measure.html#method-Measure-aggregate'><code>mlr3::Measure$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format"><a href='../../mlr3/html/Measure.html#method-Measure-format'><code>mlr3::Measure$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help"><a href='../../mlr3/html/Measure.html#method-Measure-help'><code>mlr3::Measure$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="obs_loss"><a href='../../mlr3/html/Measure.html#method-Measure-obs_loss'><code>mlr3::Measure$obs_loss()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print"><a href='../../mlr3/html/Measure.html#method-Measure-print'><code>mlr3::Measure$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score"><a href='../../mlr3/html/Measure.html#method-Measure-score'><code>mlr3::Measure$score()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureClassifConfidenceECE-new"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureClassifConfidenceECE-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{MeasureClassifConfidenceECE} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureClassifConfidenceECE$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureClassifConfidenceECE-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureClassifConfidenceECE-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureClassifConfidenceECE$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
