% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureClassifCoxIntercept.R
\name{MeasureClassifCoxIntercept}
\alias{MeasureClassifCoxIntercept}
\alias{mlr_measures_classif.cox_intercept}
\title{Cox calibration intercept}
\description{
Fits a logistic regression model with the log-odds of the predicted
probabilities as the independent variable and returns the fitted intercept.
Perfect calibration corresponds to an intercept of 0.
No single optimization direction (\code{minimize = NA}); use as a diagnostic, not a tuning objective.
}
\examples{
if (mlr3misc::require_namespaces("mlbench", quietly = TRUE) &&
  mlr3misc::require_namespaces("mlr3learners", quietly = TRUE) &&
  mlr3misc::require_namespaces("rpart", quietly = TRUE)) {
  data("Sonar", package = "mlbench")
  task = mlr3::as_task_classif(Sonar, target = "Class", positive = "M")
  learner = mlr3::lrn("classif.rpart", predict_type = "prob")
  learner$train(task)
  learner$predict(task)$score(mlr3::msr("classif.cox_intercept"))
}

}
\references{
Cox DR (1958).
\dQuote{Two Further Applications of a Model for Binary Regression.}
\emph{Biometrika}, \bold{45}(3/4), 562-565.
\doi{10.1093/biomet/45.3-4.562}.
}
\section{Super classes}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{\link[mlr3:MeasureClassif]{mlr3::MeasureClassif}} -> \code{MeasureClassifCoxIntercept}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeasureClassifCoxIntercept-new}{\code{MeasureClassifCoxIntercept$new()}}
\item \href{#method-MeasureClassifCoxIntercept-clone}{\code{MeasureClassifCoxIntercept$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate"><a href='../../mlr3/html/Measure.html#method-Measure-aggregate'><code>mlr3::Measure$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format"><a href='../../mlr3/html/Measure.html#method-Measure-format'><code>mlr3::Measure$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help"><a href='../../mlr3/html/Measure.html#method-Measure-help'><code>mlr3::Measure$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="obs_loss"><a href='../../mlr3/html/Measure.html#method-Measure-obs_loss'><code>mlr3::Measure$obs_loss()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print"><a href='../../mlr3/html/Measure.html#method-Measure-print'><code>mlr3::Measure$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score"><a href='../../mlr3/html/Measure.html#method-Measure-score'><code>mlr3::Measure$score()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureClassifCoxIntercept-new"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureClassifCoxIntercept-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{MeasureClassifCoxIntercept} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureClassifCoxIntercept$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureClassifCoxIntercept-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureClassifCoxIntercept-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureClassifCoxIntercept$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
