% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeasureClassifICI.R
\name{MeasureClassifICI}
\alias{MeasureClassifICI}
\alias{mlr_measures_classif.ici}
\title{Integrated calibration index}
\description{
Calculates the integrated calibration index (ICI) for binary classification
predictions.
A LOESS smoother is used below 1,000 observations, and a cubic regression spline is used otherwise.
Returns \code{NaN} when the calibration sample cannot identify a smooth calibration curve:
when only one outcome is observed, when there are fewer than three unique predicted probabilities,
or when the smoother fails or produces non-finite fitted values.
}
\examples{
if (mlr3misc::require_namespaces("mlbench", quietly = TRUE) &&
  mlr3misc::require_namespaces("mlr3learners", quietly = TRUE) &&
  mlr3misc::require_namespaces("rpart", quietly = TRUE)) {
  data("Sonar", package = "mlbench")
  task = mlr3::as_task_classif(Sonar, target = "Class", positive = "M")
  learner = mlr3::lrn("classif.rpart", predict_type = "prob")
  learner$train(task)
  learner$predict(task)$score(mlr3::msr("classif.ici"))
}

}
\references{
Austin PC, Steyerberg EW (2019).
\dQuote{The Integrated Calibration Index (ICI) and Related Metrics for Quantifying the Calibration of Logistic Regression Models.}
\emph{Statistics in Medicine}, \bold{38}(21), 4051-4065.
\doi{10.1002/sim.8281}.

Weissman G (2024).
\emph{gmish: Miscellaneous Functions for Predictive Modeling}.
R package version 0.1.0, commit b1ef2af, \url{https://github.com/gweissman/gmish/tree/b1ef2afef76ec2f232f5ad1d3697f31b39377204}.
}
\section{Super classes}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{\link[mlr3:MeasureClassif]{mlr3::MeasureClassif}} -> \code{MeasureClassifICI}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MeasureClassifICI-new}{\code{MeasureClassifICI$new()}}
\item \href{#method-MeasureClassifICI-clone}{\code{MeasureClassifICI$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate"><a href='../../mlr3/html/Measure.html#method-Measure-aggregate'><code>mlr3::Measure$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format"><a href='../../mlr3/html/Measure.html#method-Measure-format'><code>mlr3::Measure$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help"><a href='../../mlr3/html/Measure.html#method-Measure-help'><code>mlr3::Measure$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="obs_loss"><a href='../../mlr3/html/Measure.html#method-Measure-obs_loss'><code>mlr3::Measure$obs_loss()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print"><a href='../../mlr3/html/Measure.html#method-Measure-print'><code>mlr3::Measure$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score"><a href='../../mlr3/html/Measure.html#method-Measure-score'><code>mlr3::Measure$score()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureClassifICI-new"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureClassifICI-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{MeasureClassifICI} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureClassifICI$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MeasureClassifICI-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MeasureClassifICI-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MeasureClassifICI$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
