% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibratorBeta.R
\name{mlr3_calibrators_beta}
\alias{mlr3_calibrators_beta}
\alias{CalibratorBeta}
\title{Beta calibrator}
\description{
Beta calibration (Kull et al. 2017) for binary probabilities.
Fits the requested parameterization with \code{stats::glm} on transformed features.
Coefficients use the canonical beta-map signs for the \code{log(p)} and \code{-log(1 - p)} terms.
}
\references{
Kull M, Silva Filho T, Flach P (2017).
\dQuote{Beta Calibration: A Well-Founded and Easily Implemented Improvement on Logistic Calibration for Binary Classifiers.}
In \emph{Proceedings of the 20th International Conference on Artificial Intelligence and Statistics}, volume 54, 623-631.
\url{https://proceedings.mlr.press/v54/kull17a.html}.

Silva Filho TM, Kull M (2017).
\emph{betacal: Beta Calibration}.
R package version 0.1.0, \url{https://CRAN.R-project.org/package=betacal}.
}
\section{Super class}{
\code{\link[mlr3calibration:Calibrator]{mlr3calibration::Calibrator}} -> \code{CalibratorBeta}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CalibratorBeta-new}{\code{CalibratorBeta$new()}}
\item \href{#method-CalibratorBeta-clone}{\code{CalibratorBeta$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3calibration" data-topic="Calibrator" data-id="predict"><a href='../../mlr3calibration/html/Calibrator.html#method-Calibrator-predict'><code>mlr3calibration::Calibrator$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3calibration" data-topic="Calibrator" data-id="reset"><a href='../../mlr3calibration/html/Calibrator.html#method-Calibrator-reset'><code>mlr3calibration::Calibrator$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3calibration" data-topic="Calibrator" data-id="train"><a href='../../mlr3calibration/html/Calibrator.html#method-Calibrator-train'><code>mlr3calibration::Calibrator$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalibratorBeta-new"></a>}}
\if{latex}{\out{\hypertarget{method-CalibratorBeta-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalibratorBeta$new(parameters = "abm")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parameters}}{\code{character(1)}
Parameterization for beta calibration.
\code{"abm"} (default) is the full three-parameter family; \code{"ab"} fixes the midpoint at 0.5.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalibratorBeta-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CalibratorBeta-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalibratorBeta$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
