% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibratorIsotonic.R
\name{mlr3_calibrators_isotonic}
\alias{mlr3_calibrators_isotonic}
\alias{CalibratorIsotonic}
\title{Isotonic calibrator}
\description{
Calibration via isotonic regression.
Fitted values are linearly interpolated between observed score values and extended as constants
beyond the training range.
}
\references{
Zadrozny B, Elkan C (2002).
\dQuote{Transforming Classifier Scores into Accurate Multiclass Probability Estimates.}
In \emph{Proceedings of the 8th ACM SIGKDD International Conference on Knowledge Discovery and Data Mining}, 694-699.
\doi{10.1145/775047.775151}.

Niculescu-Mizil A, Caruana R (2005).
\dQuote{Predicting Good Probabilities with Supervised Learning.}
In \emph{Proceedings of the 22nd International Conference on Machine Learning}, 625-632.
\doi{10.1145/1102351.1102430}.
}
\section{Super class}{
\code{\link[mlr3calibration:Calibrator]{mlr3calibration::Calibrator}} -> \code{CalibratorIsotonic}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CalibratorIsotonic-new}{\code{CalibratorIsotonic$new()}}
\item \href{#method-CalibratorIsotonic-clone}{\code{CalibratorIsotonic$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3calibration" data-topic="Calibrator" data-id="predict"><a href='../../mlr3calibration/html/Calibrator.html#method-Calibrator-predict'><code>mlr3calibration::Calibrator$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3calibration" data-topic="Calibrator" data-id="reset"><a href='../../mlr3calibration/html/Calibrator.html#method-Calibrator-reset'><code>mlr3calibration::Calibrator$reset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3calibration" data-topic="Calibrator" data-id="train"><a href='../../mlr3calibration/html/Calibrator.html#method-Calibrator-train'><code>mlr3calibration::Calibrator$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalibratorIsotonic-new"></a>}}
\if{latex}{\out{\hypertarget{method-CalibratorIsotonic-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalibratorIsotonic$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CalibratorIsotonic-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CalibratorIsotonic-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CalibratorIsotonic$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
