% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3calibration-package.R
\docType{package}
\name{mlr3calibration-package}
\alias{mlr3calibration}
\alias{mlr3calibration-package}
\title{mlr3calibration package}
\description{
Probability calibration in the 'mlr3' ecosystem. Provides a unified calibration PipeOp for 'mlr3pipelines', calibration measures for 'mlr3', and reliability plots for comparing calibrated and uncalibrated learners. Supports binary and multiclass calibration strategies such as Platt scaling, isotonic regression, beta calibration, and one-vs-rest wrappers, together with published calibration-error measures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/giuseppec/mlr3calibration}
  \item Report bugs at \url{https://github.com/giuseppec/mlr3calibration/issues}
}

}
\author{
\strong{Maintainer}: Giuseppe Casalicchio \email{giuseppe.casalicchio@stat.uni-muenchen.de} (\href{https://orcid.org/0000-0001-5324-5966}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Adrian Glos \email{adrianglos117@googlemail.com}
  \item Elisabeth Meijer \email{e.meijer@campus.lmu.de}
}

}
