% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalibrationPlot.R
\name{calibrationplot}
\alias{calibrationplot}
\title{Calibration Plot Function}
\usage{
calibrationplot(
  learners,
  task,
  bins = 11,
  smooth = FALSE,
  CI = FALSE,
  rug = FALSE
)
}
\arguments{
\item{learners}{\code{list} of \code{\link[mlr3:Learner]{Learner}}\cr
List of trained learners to be evaluated.}

\item{task}{\code{\link[mlr3:Task]{Task}}\cr
Task containing the data to be used for prediction and evaluation.}

\item{bins}{\code{integer(1)}\cr
Number of bins to use when grouping predicted probabilities. Default is \code{11}.}

\item{smooth}{\code{logical(1)}\cr
Whether to plot a smoothed calibration curve using LOESS. Default is \code{FALSE}.}

\item{CI}{\code{logical(1)}\cr
Whether to include confidence intervals when \code{smooth} is \code{TRUE}. Default is \code{FALSE}.}

\item{rug}{\code{logical(1)}\cr
Whether to add a rug plot to show individual prediction points. Default is \code{FALSE}.}
}
\value{
A \code{ggplot} object displaying the calibration plot(s).
}
\description{
Generates calibration plots, also known as reliability curves, for one or more classification learners.
This function evaluates how well the predicted probabilities from the learners are calibrated with the true outcomes.
}
\details{
For each learner, the function predicts probabilities on the given task.
The predicted probabilities are divided into bins, and within each bin,
the mean predicted probability and the mean observed outcome are calculated.
These values are then plotted to assess calibration.
}
\examples{
# Example usage of calibrationplot
library(ggplot2)
# List the learners you want to plot
lrns = list(learner_uncal, learner_cal)
# Plot the reliability curves
calibrationplot(lrns, task_test, smooth = TRUE)
}
\references{
Niculescu-Mizil, A., & Caruana, R. (2005). Predicting good probabilities with supervised learning. In Proceedings of the 22nd international conference on Machine learning (pp. 625-632).
}
\concept{Plotting Functions}
