% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ici.R
\name{ici}
\alias{ici}
\title{Integrated Calibration Index}
\description{
Calculates the Expected Calibration Error (ECE) for classification tasks.
ECE measures the difference between predicted probabilities and actual outcomes,
providing an assessment of how well the predicted probabilities are calibrated.
}
\examples{
# Example usage
set.seed(1)
library(mlr3verse)

# Load the task
data("Sonar", package = "mlbench")
task = as_task_classif(Sonar, target = "Class", positive = "M")
splits = partition(task)
task_train = task$clone()$filter(splits$train)
task_test = task$clone()$filter(splits$test)

# Initialize the base learner
learner_uncal <- lrn("classif.ranger", predict_type = "prob")

# Initialize the calibrated learner
rsmp <- rsmp("cv", folds = 5)
learner_cal <- as_learner(PipeOpCalibration$new(learner = learner_uncal,
                                                method = "platt",
                                                rsmp = rsmp))

# Set ID's for the learners
learner_cal$id <- "Calibrated Learner"

# Train the calibrated learner
learner_cal$train(task_train)

# Predict the learner
prediction <- learner_cal$predict(task_test)

# Calculate the ECE
ici <- prediction$score(ici$new())

}
\references{
Weissman G (2024). \emph{gmish: Miscellaneous Functions for Predictive Modeling}. R package version 0.1.0, commit b1ef2afef76ec2f232f5ad1d3697f31b39377204, \url{https://github.com/gweissman/gmish}
}
\section{Super classes}{
\code{\link[mlr3:Measure]{mlr3::Measure}} -> \code{\link[mlr3:MeasureClassif]{mlr3::MeasureClassif}} -> \code{ici}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ici-new}{\code{ici$new()}}
\item \href{#method-ici-clone}{\code{ici$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="aggregate"><a href='../../mlr3/html/Measure.html#method-Measure-aggregate'><code>mlr3::Measure$aggregate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="format"><a href='../../mlr3/html/Measure.html#method-Measure-format'><code>mlr3::Measure$format()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="help"><a href='../../mlr3/html/Measure.html#method-Measure-help'><code>mlr3::Measure$help()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="print"><a href='../../mlr3/html/Measure.html#method-Measure-print'><code>mlr3::Measure$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="mlr3" data-topic="Measure" data-id="score"><a href='../../mlr3/html/Measure.html#method-Measure-score'><code>mlr3::Measure$score()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ici-new"></a>}}
\if{latex}{\out{\hypertarget{method-ici-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{ici} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ici$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ici-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ici-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ici$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
